+++ ghedt/examples/FromFile/create_near_square_input_file.py
@@ -5,7 +5,6 @@ import ghedt as dt
 import ghedt.PLAT as PLAT
 import ghedt.PLAT.pygfunction as gt
 import pandas as pd
-from time import time as clock
 
 
 def main():
@@ -18,17 +17,34 @@ def main():
 
     # Pipe dimensions
     # ---------------
+    # Single and Multiple U-tubes
     r_out = 26.67 / 1000. / 2.  # Pipe outer radius (m)
     r_in = 21.6 / 1000. / 2.  # Pipe inner radius (m)
     s = 32.3 / 1000.  # Inner-tube to inner-tube Shank spacing (m)
     epsilon = 1.0e-6  # Pipe roughness (m)
+    # Coaxial tube
+    r_in_in = 44.2/1000./2.
+    r_in_out = 50./1000./2.
+    # Outer pipe radii
+    r_out_in = 97.4 / 1000. / 2.
+    r_out_out = 110./1000./2.
+    # Pipe radii
+    # Note: This convention is different from pygfunction
+    r_inner = [r_in_in, r_in_out]   # The radii of the inner pipe from in to out
+    r_outer = [r_out_in, r_out_out] # The radii of the outer pipe from in to out
 
     # Pipe positions
     # --------------
     # Single U-tube [(x_in, y_in), (x_out, y_out)]
-    pos = PLAT.media.Pipe.place_pipes(s, r_out, 1)
+    pos_single = PLAT.media.Pipe.place_pipes(s, r_out, 1)
     # Single U-tube BHE object
-    bhe_object = PLAT.borehole_heat_exchangers.SingleUTube
+    single_u_tube = PLAT.borehole_heat_exchangers.SingleUTube
+    # Double U-tube
+    pos_double = PLAT.media.Pipe.place_pipes(s, r_out, 2)
+    double_u_tube = PLAT.borehole_heat_exchangers.MultipleUTube
+    # Coaxial tube
+    pos_coaxial = (0, 0)
+    coaxial_tube = PLAT.borehole_heat_exchangers.CoaxialPipe
 
     # Thermal conductivities
     # ----------------------
@@ -45,7 +61,12 @@ def main():
     # Thermal properties
     # ------------------
     # Pipe
-    pipe = PLAT.media.Pipe(pos, r_in, r_out, s, epsilon, k_p, rhoCp_p)
+    pipe_single = \
+        PLAT.media.Pipe(pos_single, r_in, r_out, s, epsilon, k_p, rhoCp_p)
+    pipe_double = \
+        PLAT.media.Pipe(pos_double, r_in, r_out, s, epsilon, k_p, rhoCp_p)
+    pipe_coaxial = \
+        PLAT.media.Pipe(pos_coaxial, r_inner, r_outer, 0, epsilon, k_p, rhoCp_p)
     # Soil
     ugt = 18.3  # Undisturbed ground temperature (degrees Celsius)
     soil = PLAT.media.Soil(k_s, rhoCp_s, ugt)
@@ -60,7 +81,7 @@ def main():
     fluid = gt.media.Fluid(mixer=mixer, percent=percent)
 
     # Fluid properties
-    V_flow_borehole = 0.2  # System volumetric flow rate (L/s)
+    V_flow_borehole = 0.2  # Borehole volumetric flow rate (L/s)
 
     # Define a borehole
     borehole = gt.boreholes.Borehole(H, D, r_b, x=0., y=0.)
@@ -85,7 +106,7 @@ def main():
     # -----------------------
     # read in the csv file and convert the loads to a list of length 8760
     hourly_extraction: dict = \
-        pd.read_csv('Atlanta_Office_Building_Loads.csv').to_dict('list')
+        pd.read_csv('../Atlanta_Office_Building_Loads.csv').to_dict('list')
     # Take only the first column in the dictionary
     hourly_extraction_ground_loads: list = \
         hourly_extraction[list(hourly_extraction.keys())[0]]
@@ -97,31 +118,30 @@ def main():
     geometric_constraints = dt.media.GeometricConstraints(
         B_max_x=B, unconstrained=True)
 
-    design = dt.design.Design(
-        V_flow_borehole, borehole, bhe_object, fluid, pipe, grout, soil,
-        sim_params, geometric_constraints, coordinates_domain,
-        hourly_extraction_ground_loads, routine='near-square', flow='borehole')
+    # Note: Flow functionality is currently only on a borehole basis. Future
+    # development will include the ability to change the flow rate to be on a
+    # system flow rate basis.
+    design_single_u_tube = dt.design.Design(
+        V_flow_borehole, borehole, single_u_tube, fluid, pipe_single, grout,
+        soil, sim_params, geometric_constraints, coordinates_domain,
+        hourly_extraction_ground_loads, routine='near-square')
 
-    print('Beginning bisection search to select a configuration.')
-    tic = clock()
-    bisection_search = design.find_design()
-    toc = clock()
-    print('Time to perform bisection search: {0:.2f} seconds'.format(toc - tic))
+    # Output the design interface object to a json file so it can be reused
+    design_single_u_tube.create_input_file(file_name='ghedt_input')
 
-    print('Number of boreholes: {}'.
-          format(len(bisection_search.selected_coordinates)))
+    design_double_u_tube = dt.design.Design(
+        V_flow_borehole, borehole, double_u_tube, fluid, pipe_double, grout,
+        soil, sim_params, geometric_constraints, coordinates_domain,
+        hourly_extraction_ground_loads, routine='near-square')
 
-    # Perform sizing in between the min and max bounds
-    tic = clock()
-    ghe = bisection_search.ghe
-    ghe.compute_g_functions()
+    design_double_u_tube.create_input_file(file_name='double_u_tube')
 
-    ghe.size(method='hybrid')
-    toc = clock()
-    print('Time to compute g-functions and size: {0:.2f} '
-          'seconds'.format(toc - tic))
+    design_coaxial_u_tube = dt.design.Design(
+        V_flow_borehole, borehole, coaxial_tube, fluid, pipe_coaxial, grout,
+        soil, sim_params, geometric_constraints, coordinates_domain,
+        hourly_extraction_ground_loads, routine='near-square')
 
-    print('Sized height of boreholes: {0:.2f} m'.format(ghe.bhe.b.H))
+    design_coaxial_u_tube.create_input_file(file_name='coaxial_tube')
 
 
 if __name__ == '__main__':